Add fake OAuth server
Add some extra paths to emulate an OAuth server for testing.

We use oauth2-server for the basic behaviour and express-oauth-server
to expose it. The redirect Uri and access and refresh token lifetimes
are configurable via command line args.

There is no authentication, the server pretends the user always grants
access to the requester.

Tokens are stored in memory so restarting the server revokes all tokens.

The /requiresAuthentication path is only accessible with a valid token,
the existing paths do not require it to avoid changing the existing
behaviour.