Fix GET and POST "/" request body parsing
The addition of body-parser middleware broke the GET and POST handlers
for the root as it was handling the on data and on end events.

This commit cleans up those paths and also adds parsing via body-parser
for text, XML and simple form data. This also means that the header
and data from the request is no longer serialized as a string inside the
response payload but an object and can be better processed on the
client.